requirements.txt
streamlit
pyyaml
python-dotenv
pymupdf
sentence-transformers
faiss-cpu
numpy
scikit-learn
tqdm
requests
